security-pipeline/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.6.1
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.23.2
aiofiles==23.2.1